Move actions to backend folder, triggering renames/import modification.
