+++ src/app/(pages)/(auth)/page.tsx
@@ -1,6 +1,6 @@
 import AuthForm from "@/app/(pages)/(auth)/components/AuthForm";
 import { redirect } from "next/navigation";
-import getSession from "@/api/actions/getSession";
+import getSession from "@/backend/actions/getSession";
 
 export default async function page() {
   const sessionData = await getSession();
