+++ src/app/api/conversations/route.ts
@@ -1,4 +1,4 @@
-import getCurrentProfile from "@/api/actions/getCurrentProfile";
+import getCurrentProfile from "@/backend/actions/getCurrentProfile";
 import { NextResponse } from "next/server";
 import prisma from "@/server/prisma";
