+++ src/components/SignOut.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { logout } from "../api/auth/actions";
+import { logout } from "../backend/auth/actions";
 
 export default function SignOut() {
   return (
