+++ src/hooks/useRoutes.ts
@@ -3,7 +3,7 @@ import { usePathname } from "next/navigation";
 import { HiChat } from "react-icons/hi";
 import { HiArrowLeftOnRectangle, HiUsers } from "react-icons/hi2";
 import useConversation from "./useConversation";
-import { logout } from "@/api/auth/actions";
+import { logout } from "@/backend/auth/actions";
 
 export default function useRoutes() {
   const pathname = usePathname();
@@ -31,7 +31,7 @@ export default function useRoutes() {
         label: "Logout",
         href: "#",
         onClick: () => logout(),
-        icon: HiArrowLeftOnRectangle
+        icon: HiArrowLeftOnRectangle,
       },
     ],
     [pathname, conversationId]
